fix: migration orders

+++ database/migrations/20220207092414_insert_users.ts
@@ -1,16 +1,18 @@
 import { hash } from 'bcrypt';
 import { Knex } from 'knex';
 
+import { UserModel } from '@/domain/models';
+import { envConfig } from '@/main/config';
+
 const tableName = 'users';
-const users = [
+const users: (Omit<UserModel, 'password' | 'roles' | 'createdAt'> & {
+  createdAt: string;
+  roles: string;
+})[] = [
   {
     id: '00000000-0000-4000-8000-000000000001',
     createUserId: '00000000-0000-4000-8000-000000000001',
-    updateUserId: null,
-    deleteUserId: null,
     createdAt: new Date().toISOString(),
-    updatedAt: null,
-    deletedAt: null,
     name: 'Admin',
     email: 'admin@email.com',
     roles: JSON.stringify(['admin']),
@@ -18,11 +20,7 @@ const users = [
   {
     id: '00000000-0000-4000-8000-000000000002',
     createUserId: '00000000-0000-4000-8000-000000000001',
-    updateUserId: null,
-    deleteUserId: null,
     createdAt: new Date().toISOString(),
-    updatedAt: null,
-    deletedAt: null,
     name: 'Moderator',
     email: 'moderator@email.com',
     roles: JSON.stringify(['moderator']),
@@ -30,11 +28,7 @@ const users = [
   {
     id: '00000000-0000-4000-8000-000000000003',
     createUserId: '00000000-0000-4000-8000-000000000001',
-    updateUserId: null,
-    deleteUserId: null,
     createdAt: new Date().toISOString(),
-    updatedAt: null,
-    deletedAt: null,
     name: 'Normal',
     email: 'normal@email.com',
     roles: JSON.stringify([]),
@@ -42,12 +36,16 @@ const users = [
 ];
 
 export async function up(knex: Knex): Promise<void> {
+  if (envConfig.nodeEnv === 'production') return;
+
   const password = await hash('Password@123', 12);
 
   await knex.table(tableName).insert(users.map((user) => ({ ...user, password })));
 }
 
 export async function down(knex: Knex): Promise<void> {
+  if (envConfig.nodeEnv === 'production') return;
+
   const userIds = users.map((user) => user.id);
 
   await knex.table(tableName).whereIn('id', userIds).delete();
